requirements.txt
streamlit>=1.31
pandas>=2.0
numpy>=1.24
plotly>=5.18
matplotlib>=3.7
python-dotenv>=1.0
supabase>=2.0
